server/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
pyjwt
bcrypt
sse-starlette
